Show tracks for current artist in second pane
As the user scrolls through the list of artists in the left panel,
the list of tracks matching that artist is show in the right.
This is a first step towards an interface for practically accessing
track information. Ideally albums for an artist should also be
shown and the track list divided up into albums.